feat: add helper rpc command to generate complex token for specific app
