+++ go.mod
@@ -2,9 +2,9 @@ module github.com/vincenzopalazzo/bender
 
 go 1.18
 
-require github.com/vincenzopalazzo/cln4go/plugin v0.0.0-20220812203530-a59f9a465f10
+require github.com/vincenzopalazzo/cln4go/plugin v0.0.0-20220814201941-66f960797315
 
-require github.com/vincenzopalazzo/cln4go v0.0.1-beta.1.0.20220814145435-6f91fa93dfa4 // indirect
+require github.com/vincenzopalazzo/cln4go v0.0.1-beta.1.0.20220814212832-655f65328877 // indirect
 
 require (
 	github.com/gin-contrib/sse v0.1.0 // indirect
@@ -20,12 +20,12 @@ require (
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/pelletier/go-toml/v2 v2.0.2 // indirect
 	github.com/ugorji/go/codec v1.2.7 // indirect
-	github.com/vincenzopalazzo/cln4go/client v0.0.0-20220802085837-fb816fecb890
-	github.com/vincenzopalazzo/cln4go/comm v0.0.0-20220802085837-fb816fecb890 // indirect
+	github.com/vincenzopalazzo/cln4go/client v0.0.0-20220814201941-66f960797315
+	github.com/vincenzopalazzo/cln4go/comm v0.0.0-20220814201941-66f960797315 // indirect
 	golang.org/x/crypto v0.0.0-20220722155217-630584e8d5aa // indirect
-	golang.org/x/net v0.0.0-20220805013720-a33c5aa5df48 // indirect
-	golang.org/x/sync v0.0.0-20220722155255-886fb9371eb4
-	golang.org/x/sys v0.0.0-20220804214406-8e32c043e418 // indirect
+	golang.org/x/net v0.0.0-20220812174116-3211cb980234 // indirect
+	golang.org/x/sync v0.0.0-20220722155255-886fb9371eb4 // indirect
+	golang.org/x/sys v0.0.0-20220811171246-fbc7d0a398ab // indirect
 	golang.org/x/text v0.3.7 // indirect
 	google.golang.org/protobuf v1.28.1 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
